Do not assume everything starting with '/' is a command
People are often unpleasantly surprised when they type a message starting with
a slash and when they send it, it just disappears without a trace.
Few people use commands in chats and aren't aware commands exist.

We should not treat everything beginning with a '/' as a command and make it
disappear.  Instead let's first make sure the command was recognized and only
then prevent sending it out as a message. If it's not recognized as a command,
assume the user wanted to send a normal message and the slash is a part of it.

Do the change for both chat and groupchat.