Improve the /google command slightly
Start matching commands using regular expressions for better flexibility.
Improve the /google command to allow opening the browser with or without a
string to search on Google. Notice that it now does not recognize e.g.
/googlefight as /google, which is a good thing.

Update both chat and groupchat.

BTW, the whole /google command handling would be better off in a plugin or at
least a separate function. And the string passed to ?q=... should be properly
escaped.  Maybe later.