abstractchatwidget: look for long smiley strings first
In some smiley themes there are cases when a smiley string is fully contained
in another string, e.g. ':)' and ':)]' in Pidgin smileys. As the iteration
order of a dictionary is undefined, it may be impossible for the user to get
the replacement of the longer smiley.  When replacing smiley strings with
pictures, we have to replace the longer strings first.
This bug was noticed by satanista_linux in fedora@conf.