storage: fix a popular traceback 'table caps has no identity column'
Old versions of Jabbim did not have the 'identity' column in the caps table.
The intention of storage.py was to drop the caps table if identity is not
there, but the code for it is wrong. It does not realize that sqlite operations
use Deferreds. It caused lots of tracebacks in logs, making it hard to see real
problems.
This patch fixes it and does some cleanups in the area as well.