userrating: adjust the calculation to match the original deduction exactly
In the deduction a_0=1, but the first actually used value in the calculation is
a_1. Switch the order of the recurrent statements to match the program with the
deduction. The useful result of this is that now the ratings are the same as
those given by the original algorithm immediately after every renormalization.