lastxml improvements
Do not just truncate the lastxml log when the limit count of messages is
written, but rotate the file to 'lastxml.1' so that there are always some
messages recorded.

Increase the limit to 50 messages. I found it useful.

Better readability of the lastxml file (separate messages with newlines).

Do not reopen the lastxml file for every message, keep the file open.