use system-wide configobj.py with 'make install'
We provide an internal copy of configobj.py. Generally this makes most
people's life easier. Otherwise it would be one more dependency for SVN
users and another external module to add to win32 builds.

Private copies of libraries are generally frowned upon in distributions,
so for proper installations using 'make install' (and thus for Linux
packages) we should use the configobj module provided by the distro.

Add the file to installation-blacklist.

Packagers must add a dependency on python-configobj. Direct SVN users
and Win32 users are not affected by the change.