Fix Plugins menu in mainWindow.ui
The UI file was modified by hand, but there was a mistake in the definition of
the Plugins submenu. Therefore it was not possible to resaved the *.ui in the
Designer without losing the submenu.
This fixes the definition of the submenu, so the file will be editable in
Designer again.