remove 0.2 s delay between loading avatar images
After logging in with Jabbim with a cold disk cache, it takes quite a
long time before all avatars are loaded. During this time, the disk
makes unpleasant periodic noise as it loads an avatar, pauses, loads
another one, pauses, ...
With warm cache we're just delaying loading of avatars for no good
reason.
Removing the delay results in a shorter total time to load the avatars,
during which the disk is utilized continuously. With warm cache the
avatars are loaded practically instantly.
There's a risk that this would negatively impact performance of other
threads. Interactivity of the GUI was fine in my testing though.