fix nonsense avatar queries
When entering a MUC, my Jabbim generates several avatar queries to the
room's JID (with no resource). It gets "not implemented" in return which
get into to the log as unhandled errors in Deferreds.

When we receive a presence from a MUC participant, we should try to get
the avatar for the full JID, not the room's JID.

This also adds some debugging prints because the avatar cache is not
working as it should and I want to find out what's going on.