Update configobj.py to 4.6.0
Our internal configobj.py copies are outdated and buggy. This updates
them to 4.6.0 which is tested with Jabbim because it's shipped in Fedora
where the jabbim package uses it.

Long term I'd prefer to remove our copies of configobj.py and just
declare ConfigObj as a required dependency, but not yet.